requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.2.0
pymupdf>=1.23.0
faiss-cpu>=1.8.0
tiktoken>=0.7.0
boto3>=1.34.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0